Some relations should only be created if min!=max
Consider a wall along the x axis. Before, we created a relation for
minz and one for maxz, even though they both refered to the same
point.  This lead to incorrect instructions because feature tracking
assumes that we can describe a wall with minx,minz,maxz,height.
This is obviously only true if minz!=maxz.